Renamed alot of files
Also renamed functions and variables to be more ruby-like.
