+++ factory/dok_factory.rb
@@ -1,4 +1,4 @@
-require './factory/StorageFactory'
+require '../factory/storage_factory'
 
 # NOTE: Dictionary of Keys (DOK)
 class DokFactory < StorageFactory
