+++ storage/sparse_storage.rb
@@ -1,5 +1,3 @@
-
-
 class SparseStorage
 	include Test::Unit::Assertions
